Refactor RequestState
