+++ clients/src/test/java/org/apache/kafka/clients/consumer/internals/RequestStateTest.java
@@ -1,0 +1,48 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.kafka.clients.consumer.internals;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+public class RequestStateTest {
+    @Test
+    public void testRequestStateSimple() {
+        RequestState state = new RequestState(
+                100,
+                2,
+                1000,
+                0);
+
+        // ensure not permitting consecutive requests
+        assertTrue(state.canSendRequest(0));
+        state.updateLastSend(0);
+        assertFalse(state.canSendRequest(0));
+        state.updateLastFailedAttempt(35);
+        assertTrue(state.canSendRequest(135));
+        state.updateLastFailedAttempt(140);
+        assertFalse(state.canSendRequest(200));
+        // exponential backoff
+        assertTrue(state.canSendRequest(340));
+
+        // test reset
+        state.reset();
+        assertTrue(state.canSendRequest(200));
+    }
+}
